feat: add user interface for displaying profile data
The previously created thunk is dispatched in Profile.jsx. The received
date from the response is sent to the ProfileExcerpt component, which
displays it. Profile.jsx handles the status.

Of course, everything could be done through suspense, initially I did
it, but in this case, when loading, the ProfileExcerpt component tried
to render, but there were no necessary props, which led to an error and
application stop.