feat: add auto authentication when entering the site
When you enter the site, authentication is launched, if successful, the
data about the user is entered in the store and the user gets to his page.

In addition to this, an error display was also added in all cases. The
logout option is still not worked out in terms of user interface