fix: display error when needed
The changes affected mainly the reducers and thunks that were in them.
Previously, reject of thunk was called on error using the thunkAPI, but
this method was not quite correct and therefore was replaced by simply
the rejectWithValue function. In addition, it was not the message itself
that was returned as the error message, but the object.