feat: update status
When you double click on the status, the status editor will open, where
you can enter a new one, and with the blur, a dispatch thunk will occur,
which, using the api method, will post the new status to the server.

It should be noted that this option may still not be optimized, in other
words, a refactor is needed.