feat: add the ability to update the profile photo
Having entered your profile, taking into account whether you are
registered and are the owner of the profile, an input will appear below
the photo in which you can select your photo, after selecting it, time
will pass and the photo will be loaded.

At the moment, the input is not fully processed, so it is not blocked
when the file is sent.