feat: create business logic to display news
The api was changed, that is, it was divided into its sections, there is
an api for a social network, and there is an api for news, all in
separate files. Obviously, a new api and method has been added to it,
namely to get news. And also a new reducer was created in which thunk is
already processed. We dispatch thunk fetchNews and receive an array of
news in return.