refactor: improve news section code
Changes to the request have been added in case of a new one, using
CancelToken. In addition, a paginator was added, which allows you to
scroll through the pages with the news.

It also now displays more of the date in the news section, however it is
not styled as the app will be styled last.

Because this api is not completely free, and a lot of the features
require a subscription, which I don't have, the apps won't always work
100% as the requests are limited.